requirements.txt
streamlit==1.8.1
pandas==1.4.2
Pillow==9.1.0
scikit-learn==1.0.2
